backend/requirements.txt
fastapi
uvicorn
ibm-watsonx-ai
langchain-ibm
langchain
langchain-chroma  # needed for watsonx ChromaVectorStore adapter
python-dotenv
python-multipart
pypdf
ibm-watson
ibm-cloud-sdk-core
requests
bs4
chromadb
python-docx
sqlalchemy
networkx
tqdm
orjson